+++ src/main/java/com/github/cc3002/finalreality/model/weapon/AbstractWeapon.java
@@ -1,5 +1,4 @@
 package com.github.cc3002.finalreality.model.weapon;
-
 import com.github.cc3002.finalreality.model.character.player.*;
 
 /**