+++ src/main/java/com/github/cc3002/finalreality/model/weapon/Sword.java
@@ -1,8 +1,6 @@
 package com.github.cc3002.finalreality.model.weapon;
-
 import com.github.cc3002.finalreality.model.character.player.Knight;
 import com.github.cc3002.finalreality.model.character.player.Thief;
-
 import java.util.Objects;
 
 /**